# src/main/java/org/example/datahub/api/TasksApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.1.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.example.datahub.api;

import org.example.datahub.model.SuccessResponseDTO;
import org.example.datahub.model.TaskCreateResponseDTO;
import org.example.datahub.model.TaskDetailResponseDTO;
import org.example.datahub.model.TaskListResponseDTO;
import org.example.datahub.model.TaskMetadataDTO;
import org.example.datahub.model.TaskSubmissionExportResponseDTO;
import org.example.datahub.model.TaskTeacherListResponseDTO;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-12-04T09:26:02.924376840+08:00[Asia/Shanghai]")
@Validated
@Tag(name = "tasks", description = "the tasks API")
public interface TasksApi {

    /**
     * POST /tasks/{task_id}/submissions/export : Export submissions
     * Export all task submissions as file.
     *
     * @param taskId  (required)
     * @return Successful response (status code 200)
     */
    @Operation(
        operationId = "exportTaskSubmissions",
        summary = "Export submissions",
        description = "Export all task submissions as file.",
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful response", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = TaskSubmissionExportResponseDTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/tasks/{task_id}/submissions/export",
        produces = { "application/json" }
    )
    
    ResponseEntity<TaskSubmissionExportResponseDTO> exportTaskSubmissions(
        @Parameter(name = "task_id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("task_id") Long taskId
    );


    /**
     * POST /tasks/{task_id}/remind_teachers : Remind unsubmitted teachers
     * Send reminder to teachers.
     *
     * @param taskId  (required)
     * @return Successful response (status code 200)
     */
    @Operation(
        operationId = "remindUnsubmittedTeachers",
        summary = "Remind unsubmitted teachers",
        description = "Send reminder to teachers.",
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful response", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = SuccessResponseDTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/tasks/{task_id}/remind_teachers",
        produces = { "application/json" }
    )
    
    ResponseEntity<SuccessResponseDTO> remindUnsubmittedTeachers(
        @Parameter(name = "task_id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("task_id") Long taskId
    );


    /**
     * POST /tasks : Create task
     * Create a new collection task.
     *
     * @param metadata  (required)
     * @param templateFile  (required)
     * @return Successful response (status code 200)
     */
    @Operation(
        operationId = "taskCreate",
        summary = "Create task",
        description = "Create a new collection task.",
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful response", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = TaskCreateResponseDTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/tasks",
        produces = { "application/json" },
        consumes = { "multipart/form-data" }
    )
    
    ResponseEntity<TaskCreateResponseDTO> taskCreate(
        @Parameter(name = "metadata", description = "", required = true) @Valid @RequestPart(value = "metadata", required = true) TaskMetadataDTO metadata,
        @Parameter(name = "template_file", description = "", required = true) @RequestPart(value = "template_file", required = true) MultipartFile templateFile
    );


    /**
     * DELETE /tasks/{task_id} : Delete task
     * Remove a task.
     *
     * @param taskId  (required)
     * @return Successful response (status code 200)
     */
    @Operation(
        operationId = "taskDelete",
        summary = "Delete task",
        description = "Remove a task.",
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful response", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = SuccessResponseDTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/tasks/{task_id}",
        produces = { "application/json" }
    )
    
    ResponseEntity<SuccessResponseDTO> taskDelete(
        @Parameter(name = "task_id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("task_id") Long taskId
    );


    /**
     * GET /tasks/{task_id} : Get task detail
     * Query task detail.
     *
     * @param taskId  (required)
     * @return Successful response (status code 200)
     */
    @Operation(
        operationId = "taskDetail",
        summary = "Get task detail",
        description = "Query task detail.",
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful response", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = TaskDetailResponseDTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/tasks/{task_id}",
        produces = { "application/json" }
    )
    
    ResponseEntity<TaskDetailResponseDTO> taskDetail(
        @Parameter(name = "task_id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("task_id") Long taskId
    );


    /**
     * GET /tasks : Get task list
     * Query list of tasks.
     *
     * @param pageNum  (optional)
     * @param pageSize  (optional)
     * @param status  (optional)
     * @param deptIds  (optional)
     * @param orderBy  (optional)
     * @param orderDirection  (optional)
     * @return Successful response (status code 200)
     */
    @Operation(
        operationId = "taskList",
        summary = "Get task list",
        description = "Query list of tasks.",
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful response", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = TaskListResponseDTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/tasks",
        produces = { "application/json" }
    )
    
    ResponseEntity<TaskListResponseDTO> taskList(
        @Parameter(name = "page_num", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page_num", required = false) Integer pageNum,
        @Parameter(name = "page_size", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page_size", required = false) Integer pageSize,
        @Parameter(name = "status", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "status", required = false) String status,
        @Parameter(name = "dept_ids", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "dept_ids", required = false) List<Long> deptIds,
        @Parameter(name = "order_by", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "order_by", required = false) String orderBy,
        @Parameter(name = "order_direction", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "order_direction", required = false) String orderDirection
    );


    /**
     * GET /tasks/{task_id}/teachers : Get teacher list by task
     * Query task teacher list with submission status.
     *
     * @param taskId  (required)
     * @param pageNum  (optional)
     * @param pageSize  (optional)
     * @param status  (optional)
     * @return Successful response (status code 200)
     */
    @Operation(
        operationId = "taskTeacherList",
        summary = "Get teacher list by task",
        description = "Query task teacher list with submission status.",
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful response", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = TaskTeacherListResponseDTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/tasks/{task_id}/teachers",
        produces = { "application/json" }
    )
    
    ResponseEntity<TaskTeacherListResponseDTO> taskTeacherList(
        @Parameter(name = "task_id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("task_id") Long taskId,
        @Parameter(name = "page_num", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page_num", required = false) Integer pageNum,
        @Parameter(name = "page_size", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page_size", required = false) Integer pageSize,
        @Parameter(name = "status", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "status", required = false) String status
    );


    /**
     * PUT /tasks/{task_id} : Update task
     * Update task information.
     *
     * @param taskId  (required)
     * @param metadata  (required)
     * @param templateFile  (optional)
     * @return Successful response (status code 200)
     */
    @Operation(
        operationId = "taskUpdate",
        summary = "Update task",
        description = "Update task information.",
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful response", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = SuccessResponseDTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/tasks/{task_id}",
        produces = { "application/json" },
        consumes = { "multipart/form-data" }
    )
    
    ResponseEntity<SuccessResponseDTO> taskUpdate(
        @Parameter(name = "task_id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("task_id") Long taskId,
        @Parameter(name = "metadata", description = "", required = true) @Valid @RequestPart(value = "metadata", required = true) TaskMetadataDTO metadata,
        @Parameter(name = "template_file", description = "") @RequestPart(value = "template_file", required = false) MultipartFile templateFile
    );

}