# src/main/resources/openapi/api.yml
openapi: 3.0.3
info:
  title: Research Collection System API
  version: 1.0.0
  description: API specification for research task collection system.
servers:
  - url: /api
paths:
...
  /tasks:
    post:
      operationId: taskCreate
      summary: Create task
      description: Create a new collection task.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                metadata:
                  $ref: "#/components/schemas/TaskMetadata"
                template_file:
                  type: string
                  format: binary
              required:
                - metadata
                - template_file
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskCreateResponse"

    get:
      operationId: taskList
      summary: Get task list
      description: Query list of tasks.
      parameters:
        - in: query
          name: page_num
          schema: { type: integer }
        - in: query
          name: page_size
          schema: { type: integer }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: dept_ids
          schema:
            type: array
            items:
              type: integer
              format: int64
        - in: query
          name: order_by
          schema: { type: string }
        - in: query
          name: order_direction
          schema: { type: string }
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskListResponse"
  /tasks/{task_id}:
    get:
      operationId: taskDetail
      summary: Get task detail
      description: Query task detail.
      parameters:
        - in: path
          name: task_id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskDetailResponse"

    put:
      operationId: taskUpdate
      summary: Update task
      description: Update task information.
      parameters:
        - in: path
          name: task_id
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                metadata:
                  $ref: "#/components/schemas/TaskMetadata"
                template_file:
                  type: string
                  format: binary
              required:
                - metadata
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"


    delete:
      operationId: taskDelete
      summary: Delete task
      description: Remove a task.
      parameters:
        - in: path
          name: task_id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /tasks/{task_id}/teachers:
    get:
      operationId: taskTeacherList
      summary: Get teacher list by task
      description: Query task teacher list with submission status.
      parameters:
        - in: path
          name: task_id
          required: true
          schema: { type: integer, format: int64 }
        - in: query
          name: page_num
          schema: { type: integer }
        - in: query
          name: page_size
          schema: { type: integer }
        - in: query
          name: status
          schema: { type: string }
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskTeacherListResponse"
  /tasks/{task_id}/remind_teachers:
    post:
      operationId: remindUnsubmittedTeachers
      summary: Remind unsubmitted teachers
      description: Send reminder to teachers.
      parameters:
        - in: path
          name: task_id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

##### 2.2.8 Collect Submissions
#
#  + POST /tasks/{task_id}/submissions/export
#  + Response
#  ```json
#  {
#    "success": true,
#    "data": {
#      "file_id": "1234567890"
#    },
#    "message": "Export submissions successfully."
#  }
#  ```
  /tasks/{task_id}/submissions/export:
    post:
      operationId: exportTaskSubmissions
      summary: Export submissions
      description: Export all task submissions as file.
      parameters:
        - in: path
          name: task_id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskSubmissionExportResponse"
...
components:
  schemas:
...
    TaskMetadata:
      type: object
      properties:
        task_name: { type: string }
        description: { type: string }
        deadline: { type: string, format: date-time }
        dept_id: { type: integer, format: int64 }
      required: [ task_name, deadline, dept_id ]
    TaskCreateResponse:
      type: object
      properties:
        success: { type: boolean }
        data:
          type: object
          properties:
            task_id: { type: integer, format: int64 }
        message: { type: string }
    TaskListResponse:
      type: object
      properties:
        success: { type: boolean }
        data:
          type: object
          properties:
            tasks:
              type: array
              items:
                $ref: "#/components/schemas/TaskItem"
            page:
              $ref: "#/components/schemas/PageInfo"
        message: { type: string }
    TaskItem:
      type: object
      properties:
        task_id: { type: integer, format: int64 }
        task_name: { type: string }
        description: { type: string }
        template_file_id: { type: integer, format: int64 }
        deadline: { type: string, format: date-time }
        status: { type: string }
        create_time: { type: string, format: date-time }
        department:
          $ref: "#/components/schemas/DepartmentItem"
    TaskDetailResponse:
      type: object
      properties:
        success: { type: boolean }
        data:
          $ref: "#/components/schemas/TaskItem"
        message: { type: string }
    TaskTeacherListResponse:
      type: object
      properties:
        success: { type: boolean }
        data:
          type: object
          properties:
            task_id: { type: integer, format: int64 }
            teachers:
              type: array
              items:
                $ref: "#/components/schemas/TaskTeacherItem"
            page:
              $ref: "#/components/schemas/PageInfo"
        message: { type: string }
    TaskTeacherItem:
      type: object
      properties:
        teacher_id: { type: integer, format: int64 }
        teacher_name: { type: string }
        teacher_email: { type: string }
        submission:
          $ref: "#/components/schemas/SubmissionStatusItem"
    SubmissionStatusItem:
      type: object
      properties:
        status: { type: string }
        submitted_at: { type: string, format: date-time }
        attachment_file_id: { type: integer, format: int64 }
        submission_id: { type: integer, format: int64 }
    TaskSubmissionExportResponse:
      type: object
      properties:
        success: { type: boolean }
        data:
          type: object
          properties:
            file_id: { type: integer, format: int64 }
        message: { type: string }

    # ==================== DEPARTMENT ====================
...
    DepartmentItem:
      type: object
      properties:
        dept_id: { type: integer, format: int64 }
        dept_name: { type: string }

    # ==================== SHARED ====================
    PageInfo:
      type: object
      properties:
        total_pages: { type: integer }
        page_num: { type: integer }
        page_size: { type: integer }
    SuccessResponse:
      type: object
      properties:
        success: { type: boolean }
        message: { type: string }
...